Launcher profile stats (#809)
* add two new types to carry the stats

* removed unused method and usings

* add methods to helper to work out money amount, todo other stats

* add missing using

* Changed The value of CounterKey to Long, this is what it is in the client

* clean up get currencies method, added methods to get profile stats for use in the launcher

+++ Libraries/SPTarkov.Server.Core/Models/Eft/Common/Tables/BotBase.cs
@@ -568,7 +568,7 @@ public record CounterKeyValue
 {
     public HashSet<string>? Key { get; set; }
 
-    public double? Value { get; set; }
+    public long? Value { get; set; }
 }
 
 public record Aggressor
